Removing unneeded mipengine folders. Removed the common and utils modules. Moved internal modules in the mipengine module. Moved the convert_mip_type_to_python_type to the datatypes.py module.
