+++ mipengine/__init__.py
@@ -1,3 +1,4 @@
 from .datatypes import DType
+from .attrdict import AttrDict
 
-__all__ = ["DType"]
+__all__ = ["DType", "AttrDict"]
