+++ mipengine/node/monetdb_interface/common_actions.py
@@ -1,9 +1,9 @@
 from typing import List
 from typing import Union
 
-from mipengine.common.node_exceptions import TablesNotFound
-from mipengine.common.node_tasks_DTOs import ColumnInfo
-from mipengine.common.node_tasks_DTOs import TableSchema
+from mipengine.node_exceptions import TablesNotFound
+from mipengine.node_tasks_DTOs import ColumnInfo
+from mipengine.node_tasks_DTOs import TableSchema
 from mipengine.node.monetdb_interface.monet_db_connection import MonetDB
 
 MONETDB_VARCHAR_SIZE = 50
@@ -22,24 +22,6 @@ def create_table_name(
         raise TypeError(f"Table type is not acceptable: {table_type} .")
 
     return f"{table_type}_{command_id}_{context_id}_{node_id}".lower()
-
-
-def convert_mip_type_to_class_type(mip_type: str):
-    """
-    Converts MIP's types to the according class.
-    """
-    type_mapping = {
-        "int": int,
-        "real": float,
-        "text": str,
-    }
-
-    if mip_type not in type_mapping.keys():
-        raise KeyError(
-            f"MIP type '{mip_type}' cannot be converted to a python class type."
-        )
-
-    return type_mapping.get(mip_type)
 
 
 def convert_schema_to_sql_query_format(schema: TableSchema) -> str:
