+++ mipengine/node/monetdb_interface/remote_tables.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from mipengine.node import config as node_config
-from mipengine.common.node_tasks_DTOs import TableInfo
+from mipengine.node_tasks_DTOs import TableInfo
 from mipengine.node.monetdb_interface.common_actions import get_table_names
 from mipengine.node.monetdb_interface.common_actions import (
     convert_schema_to_sql_query_format,
