+++ mipengine/node/monetdb_interface/tables.py
@@ -1,6 +1,6 @@
 from typing import List, Union
 
-from mipengine.common.node_tasks_DTOs import TableInfo
+from mipengine.node_tasks_DTOs import TableInfo
 from mipengine.node.monetdb_interface import common_actions
 from mipengine.node.monetdb_interface.common_actions import (
     convert_schema_to_sql_query_format,
