+++ mipengine/node/tasks/udfs.py
@@ -6,10 +6,10 @@ from typing import Tuple
 from celery import shared_task
 
 from mipengine.node import config as node_config
-from mipengine import algorithms
 
-# from mipengine.algorithms import demo  # TODO Split the actual and testing algorithms
-from mipengine.common.node_tasks_DTOs import (
+from mipengine import algorithms  # DO NOT REMOVE, NEEDED FOR ALGORITHM IMPORT
+
+from mipengine.node_tasks_DTOs import (
     UDFArgument,
     ColumnInfo,
     TableInfo,
