+++ mipengine/resources/pathologies_metadata/dementia.json
