+++ mipengine/resources/pathologies_metadata/mentalhealth.json
