--- mipengine/utils/__init__.py
@@ -1,3 +1,0 @@
-from .attrdict import AttrDict
-
-__all__ = ["AttrDict"]
