+++ tests/integration_tests/test_merge_tables.py
@@ -2,10 +2,10 @@ import uuid
 
 import pytest
 
-from mipengine.common.node_exceptions import IncompatibleSchemasMergeException
-from mipengine.common.node_exceptions import TablesNotFound
-from mipengine.common.node_tasks_DTOs import ColumnInfo
-from mipengine.common.node_tasks_DTOs import TableSchema
+from mipengine.node_exceptions import IncompatibleSchemasMergeException
+from mipengine.node_exceptions import TablesNotFound
+from mipengine.node_tasks_DTOs import ColumnInfo
+from mipengine.node_tasks_DTOs import TableSchema
 from tests.integration_tests.nodes_communication import get_celery_task_signature
 from tests.integration_tests.nodes_communication import get_celery_app
