+++ tests/integration_tests/test_tables.py
@@ -1,9 +1,9 @@
 import pytest
 import uuid as uuid
 
-from mipengine.common.node_tasks_DTOs import ColumnInfo
-from mipengine.common.node_tasks_DTOs import TableData
-from mipengine.common.node_tasks_DTOs import TableSchema
+from mipengine.node_tasks_DTOs import ColumnInfo
+from mipengine.node_tasks_DTOs import TableData
+from mipengine.node_tasks_DTOs import TableSchema
 from tests.integration_tests.nodes_communication import get_celery_app
 from tests.integration_tests.nodes_communication import get_celery_task_signature
